Subscribed to film list endpoint
The film list endpoint now return an observable which is being observed
by the presenter. When that value changes it will notify the presenter
and update the _filmList variable. I will observe the _filmList
variable from the activity so when that updates it will update the list
in the ui.